core: adapt to datascript Int64 values (pin e439cba)
Datascript upstream c39cad1 changed value Int of int to Int64 of int64.
Convert at the boundary: chat domain types stay int, wrapped via
Int64.of_int/Int64.to_int; test literals get the L suffix; sync codec
emits Transit.Int64 unconditionally like upstream. file/created-at and
file/last-modified-at are written as Int64 rather than Instant since
upstream now demotes Instant to Int64 for attrs without
:db.type/instant on storage read.

+++ shared/test/logseq_chat/e2e_seed_data_test.ml
@@ -245,7 +245,7 @@ let composer_reset_preserves_built_ins_and_removes_user_content () =
           ("block/uuid", Ds.One_value (Ds.Uuid "e2e-stale-page"));
           ("block/name", Ds.One_value (Ds.String "stale journal"));
           ("block/title", Ds.One_value (Ds.String "Stale Journal"));
-          ("block/journal-day", Ds.One_value (Ds.Int 20260827));
+          ("block/journal-day", Ds.One_value (Ds.Int64 20260827L));
         ];
       entity None
         [
